apps/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlmodel==0.0.24
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-multipart==0.0.20
httpx==0.28.1
